+++ src/main/resources/templates/qna/show.html
@@ -19,7 +19,7 @@
                         <div class="article-header-text">
                             <a href="/users/92/kimmunsu" class="article-author-name">{{question.writer}}</a>
                             <a href="/questions/413" class="article-header-time" title="퍼머링크">
-                                {{question.currentTime}}
+                                {{question.currentDateTime}}
                                 <i class="icon-link"></i>
                             </a>
                         </div>
